// src/app/shared/models/enum.ts
export const enum VERSION_1_DETAILS_LOOKUP {
  LINES_OF_CODE = '1,900',
  HOURS_WORKED = '50 hrs',
  SIZE = '2mb',
  HTML_PERCENTAGE = 58.2,
  CSS_PERCENTAGE = 40,
  SCSS_PERCENTAGE = 0,
  JAVASCRIPT_PERCENTAGE = 1.8,
  TYPESCRIPT_PERCENTAGE = 0,
}

export const enum VERSION_2_DETAILS_LOOKUP {
  LINES_OF_CODE = '6,400,000',
  HOURS_WORKED = '80 hrs',
  SIZE = '7.34mb',
  HTML_PERCENTAGE = 43.7,
  CSS_PERCENTAGE = 30.1,
  SCSS_PERCENTAGE = 0,
  JAVASCRIPT_PERCENTAGE = 2.1,
  TYPESCRIPT_PERCENTAGE = 24.1
}


export const enum VERSION_3_DETAILS_LOOKUP {
  LINES_OF_CODE = '5,800,000',
  HOURS_WORKED = '100 hrs',
  SIZE = '15mb',
  HTML_PERCENTAGE = 35.8,
  CSS_PERCENTAGE = 50.9,
  SCSS_PERCENTAGE = 0,
  JAVASCRIPT_PERCENTAGE = 1.0,
  TYPESCRIPT_PERCENTAGE = 12.3
}


export const enum VERSION_4_DETAILS_LOOKUP {
  LINES_OF_CODE = '4,600,000',
  HOURS_WORKED = '50 hrs',
  SIZE = '18.8mb',
  HTML_PERCENTAGE = 33.4,
  CSS_PERCENTAGE = 5.8,
  SCSS_PERCENTAGE = 47.5,
  JAVASCRIPT_PERCENTAGE = 1.4,
  TYPESCRIPT_PERCENTAGE = 11.9
}


export const enum VERSION_5_DETAILS_LOOKUP {
  LINES_OF_CODE = '38,700',
  HOURS_WORKED = '50 hrs',
  SIZE = '50mb',
  HTML_PERCENTAGE = 33.7,
  CSS_PERCENTAGE = 0,
  SCSS_PERCENTAGE = 44.4,
  JAVASCRIPT_PERCENTAGE = 1.1,
  TYPESCRIPT_PERCENTAGE = 20.8
}